Restore the entire client suite
